requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf_spectacular
drf_extensions